Fixes adjustedMaxMinValues logic when max 0
Addressed issue where if a series had a set of data values that had a
max of 0, the scale returned by adjustedMaxMinValues would have a
correctly set minValues property but an incorrectly set maxValues
property that was always 1.

This error was caused by vestigial logic in adjustedMaxMinValues that
should have been removed when the new axis/distortion logic was
implemented.

Unit tests for adjustedMaxMinValues and a visual verification graph
have been added to Elroi Qunit tests.