adjustedMaxMinValues avoids identical max/min pair
The adjustedMaxMinValues method is responsible for determining the
minimum and maximum values for the y-axis of an Elroi graph.   It is
possible with certain data sets and configuration to end up with the
same maximum and minumum values being returned for a series by the
maxValues and minValues helper methods.

Previously, in this case, because the graph had a range of 0, a number
of downstream calculations would break.  If the matching max/min were
positive or negative an error would be thrown and the graph data would
fail to render.  If the matching max/min were 0 Elroi would get stuck
in an infinite loop while attempting to calculate what the y-axis'
labels should be, hanging the entire browser.

It is imperative the max and min value for an axis don't match.  This
makes sense for graphs in general, as otherwise you'd have a line.

If maxValues and minValues return the same result for a series the
maxValue is now bumped up by 1.  This prevents the graph from hanging or
generating errors.  It is also done in an appropriate location so that
scaling continues to work.

Thorough notes at the location of the fix are included.  Automated unit
tests and visually verifiable test cases have been added to the
collection of Qunit tests.

remove debugger